requirements.txt
llvmlite==0.41.0
numba==0.58.0
numpy==1.25.2
opencv-contrib-python==4.8.0.76
Pillow==10.0.1
scipy==1.11.2
